fix: prevent LastPass browser extension output
An on-again, off-again problem with the LP extension over the years.
Currently on-again.

The error appears in the console but doesn't get in the way otherwise:

```js
Assertion failed: Input argument is not an HTMLInputElement

getFormProfile		@	onloadwff.js:71
setFieldValue		@	onloadwff.js:71
formKeydownListener	@	onloadwff.js:71
```

This should stop the event from bubbling up to the LP extension.